# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API for the BONUSPLAY demo rewards platform
servers:
  - url: /api
    description: Base API path
...
paths:
  /v2/admin/requests:
    get:
      operationId: listV2AdminRequests
      tags: [v2]
      summary: High-risk administrator requests, newest first
      parameters:
        - name: status
          in: query
          schema: { type: string, enum: [pending, executed, rejected, cancelled] }
      responses:
        "200":
          description: Requests with confirmation eligibility for the viewer
          content:
            application/json:
              schema: { $ref: "#/components/schemas/V2AdminRequestPage" }
        default:
          $ref: "#/components/responses/V2PointsError"
    post:
      operationId: createV2AdminRequest
      tags: [v2]
      summary: Record a high-risk action for later confirmation
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/V2AdminRequestInput" }
      responses:
        "202":
          description: Recorded or identical replay
          content:
            application/json:
              schema: { $ref: "#/components/schemas/V2AdminRequest" }
        default:
          $ref: "#/components/responses/V2PointsError"
  /v2/admin/requests/{requestId}/confirm:
    post:
      operationId: confirmV2AdminRequest
      tags: [v2]
      summary: Execute a pending request after cooldown; requires fresh MFA
      parameters:
        - $ref: "#/components/parameters/AdminRequestId"
      responses:
        "200":
          description: Executed request
          content:
            application/json:
              schema: { $ref: "#/components/schemas/V2AdminRequest" }
        default:
          $ref: "#/components/responses/V2PointsError"
  /v2/admin/requests/{requestId}/reject:
    post:
      operationId: rejectV2AdminRequest
      tags: [v2]
      summary: Cancel own or reject another administrator's pending request
      parameters:
        - $ref: "#/components/parameters/AdminRequestId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/V2AdminRejectInput" }
      responses:
        "200":
          description: Closed request
          content:
            application/json:
              schema: { $ref: "#/components/schemas/V2AdminRequest" }
        default:
          $ref: "#/components/responses/V2PointsError"
  /v2/admin/economy:
    get:
      operationId: getV2AdminEconomy
      tags: [v2]
      summary: Current economy configuration and points liability
      responses:
        "200":
          description: Economy overview
          content:
            application/json:
              schema: { $ref: "#/components/schemas/V2EconomyOverview" }
        default:
          $ref: "#/components/responses/V2PointsError"
  /v2/admin/audit:
    get:
      operationId: listV2AdminAudit
      tags: [v2]
      summary: Append-only administrative audit log, newest first
      parameters:
        - name: before
          in: query
          schema: { type: string, format: date-time }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
      responses:
        "200":
          description: Audit page
          content:
            application/json:
              schema: { $ref: "#/components/schemas/V2AuditPage" }
        default:
          $ref: "#/components/responses/V2PointsError"
...
  /v2/offers:
    get:
      operationId: listV2Offers
      tags: [v2]
      responses:
        "200":
          description: Approved offers and rollout gate
          content:
            application/json:
              schema: { $ref: "#/components/schemas/V2OfferPage" }
        default: { $ref: "#/components/responses/V2PointsError" }
  /v2/offers/{offerId}:
    get:
      operationId: getV2Offer
      tags: [v2]
      parameters:
        - { name: offerId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: Approved offer details
          content:
            application/json:
              schema: { $ref: "#/components/schemas/V2OfferDetail" }
        default: { $ref: "#/components/responses/V2PointsError" }
  /v2/offers/{offerId}/start:
    post:
      operationId: startV2Offer
      tags: [v2]
      parameters:
        - { name: offerId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/V2OfferStartInput" }
      responses:
        "200":
          description: Persisted attributed click; no points awarded
          content:
            application/json:
              schema: { $ref: "#/components/schemas/V2OfferStart" }
        default: { $ref: "#/components/responses/V2PointsError" }
  /v2/conversions:
    get:
      operationId: listV2Conversions
      tags: [v2]
      responses:
        "200":
          description: Own latest 100 conversions
          content:
            application/json:
              schema: { $ref: "#/components/schemas/V2ConversionPage" }
        default: { $ref: "#/components/responses/V2PointsError" }
  /v2/admin/offer-partners:
    get:
      operationId: listV2OfferPartners
      tags: [v2]
      responses:
        "200":
          description: Operator-configured partner identifiers; no credentials
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/V2OfferPartner" }
        default: { $ref: "#/components/responses/V2PointsError" }
  /v2/admin/offers:
    get:
      operationId: listV2AdminOffers
      tags: [v2]
      responses:
        "200":
          description: Latest 100 offers including drafts
          content:
            application/json:
              schema: { $ref: "#/components/schemas/V2OfferPage" }
        default: { $ref: "#/components/responses/V2PointsError" }
    post:
      operationId: createV2Offer
      tags: [v2]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/V2OfferInput" }
      responses:
        "201":
          description: Immutable draft, awaiting review
          content:
            application/json:
              schema: { $ref: "#/components/schemas/V2Offer" }
        default: { $ref: "#/components/responses/V2PointsError" }
  /v2/admin/offers/{offerId}/review:
    post:
      operationId: reviewV2Offer
      tags: [v2]
      parameters:
        - { name: offerId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/V2OfferReviewInput" }
      responses:
        "200":
          description: Reviewed offer
          content:
            application/json:
              schema: { $ref: "#/components/schemas/V2Offer" }
        default: { $ref: "#/components/responses/V2PointsError" }
  /v2/admin/conversions:
    get:
      operationId: listV2AdminConversions
      tags: [v2]
      responses:
        "200":
          description: Latest 100 conversions for administrator review
          content:
            application/json:
              schema: { $ref: "#/components/schemas/V2ConversionPage" }
        default: { $ref: "#/components/responses/V2PointsError" }
  /v2/admin/conversions/{conversionId}/review:
    post:
      operationId: reviewV2Conversion
      tags: [v2]
      parameters:
        - { name: conversionId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/V2ConversionReviewInput" }
      responses:
        "200":
          description: Atomic conversion and ledger decision
          content:
            application/json:
              schema: { $ref: "#/components/schemas/V2Conversion" }
        default: { $ref: "#/components/responses/V2PointsError" }
  /v2/admin/conversions/{conversionId}/reverse:
    post:
      operationId: reverseV2Conversion
      tags: [v2]
      summary: Administrator-only full reversal; preserves terminal partner evidence
      parameters:
        - { name: conversionId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/V2ConversionReversalInput" }
      responses:
        "200":
          description: Atomic compensation and immutable reversed lifecycle event; replay returns current state
          content:
            application/json:
              schema: { $ref: "#/components/schemas/V2ConversionReversalResult" }
        default: { $ref: "#/components/responses/V2PointsError" }
...
components:
  responses:
    V2PointsError:
      description: 400 invalid input; 401 login required; 403 forbidden; 404 not found; 409 replay conflict, insufficient balance or invalid transition; 429 rate limit
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
  parameters:
    AdminRequestId:
      name: requestId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  schemas:
    V2AdminRequestInput:
      type: object
      additionalProperties: false
      required: [action, payload, reason, requestKey]
      properties:
        action: { type: string, enum: [POINTS_ADJUSTMENT, POINTS_DECISION, POINTS_COMPENSATION, OFFER_APPROVAL, REWARD_APPROVAL, ORDER_DISPATCH, ECONOMY_CONFIG, MARKETING_BUDGET] }
        payload: { type: object, additionalProperties: true }
        reason: { type: string, minLength: 10, maxLength: 500 }
        requestKey: { type: string, minLength: 16, maxLength: 128, pattern: '^[A-Za-z0-9_.:-]{16,128}$' }
    V2AdminRejectInput:
      type: object
      additionalProperties: false
      required: [note]
      properties:
        note: { type: string, minLength: 10, maxLength: 500 }
    V2AdminRequest:
      type: object
      required: [id, action, payload, reason, requestedBy, requestedByName, targetAccountId, requestedAt, notBefore, status, decidedBy, decidedAt, note, result, canConfirm, blockedReason]
      properties:
        id: { type: string }
        action: { type: string, enum: [POINTS_ADJUSTMENT, POINTS_DECISION, POINTS_COMPENSATION, OFFER_APPROVAL, REWARD_APPROVAL, ORDER_DISPATCH, ECONOMY_CONFIG, MARKETING_BUDGET] }
        payload: { type: object, additionalProperties: true }
        reason: { type: string }
        requestedBy: { type: string }
        requestedByName: { type: string }
        targetAccountId: { type: [string, "null"] }
        requestedAt: { type: string }
        notBefore: { type: string }
        status: { type: string, enum: [pending, executed, rejected, cancelled] }
        decidedBy: { type: [string, "null"] }
        decidedAt: { type: [string, "null"] }
        note: { type: [string, "null"] }
        result: { type: [object, "null"], additionalProperties: true }
        canConfirm: { type: boolean }
        blockedReason: { type: [string, "null"] }
    V2AdminRequestPage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/V2AdminRequest" }
    V2EconomyConfig:
      type: object
      required: [version, pointsPerNok, defaultShareBp, maxShareBp, minMarginBp, highRiskCooldownMinutes, highValueOrderPoints, dualControl, minAccountAgeDays, minVerifiedPointsBeforeRedeem, maxRedemptionsPerDay, maxRedeemPointsPerDay, createdBy, createdAt]
      properties:
        version: { type: integer }
        pointsPerNok: { type: integer }
        defaultShareBp: { type: integer }
        maxShareBp: { type: integer }
        minMarginBp: { type: integer }
        highRiskCooldownMinutes: { type: integer }
        highValueOrderPoints: { type: integer }
        dualControl: { type: boolean }
        minAccountAgeDays: { type: integer }
        minVerifiedPointsBeforeRedeem: { type: integer }
        maxRedemptionsPerDay: { type: integer }
        maxRedeemPointsPerDay: { type: integer }
        createdBy: { type: string }
        createdAt: { type: string }
    V2Liability:
      type: object
      required: [balancePoints, reservedPoints, availablePoints, pendingCreditPoints, balanceNok, pendingCreditNok]
      properties:
        balancePoints: { type: integer }
        reservedPoints: { type: integer }
        availablePoints: { type: integer }
        pendingCreditPoints: { type: integer }
        balanceNok: { type: number }
        pendingCreditNok: { type: number }
    V2EconomyOverview:
      type: object
      required: [config, liability, admins]
      properties:
        config: { $ref: "#/components/schemas/V2EconomyConfig" }
        liability: { $ref: "#/components/schemas/V2Liability" }
        admins: { type: integer }
    V2AuditEntry:
      type: object
      required: [id, actorId, actorRole, action, entityType, entityId, metadata, createdAt]
      properties:
        id: { type: string }
        actorId: { type: string }
        actorRole: { type: string }
        action: { type: string }
        entityType: { type: string }
        entityId: { type: string }
        metadata: { type: object, additionalProperties: true }
        createdAt: { type: string }
    V2AuditPage:
      type: object
      required: [items, nextBefore]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/V2AuditEntry" }
        nextBefore: { type: [string, "null"] }
...
    V2OfferPartner:
      type: object
      required: [id, name]
      properties:
        id: { type: string }
        name: { type: string }
    V2OfferInput:
      type: object
      additionalProperties: false
      required: [partnerId, title, description, terms, points, destinationUrl, category, requirements, completionSteps, estimatedMinutes, approvalDays]
      properties:
        partnerId: { type: string, pattern: "^[a-z0-9-]{3,40}$" }
        title: { type: string, minLength: 3, maxLength: 120 }
        description: { type: string, minLength: 10, maxLength: 2000 }
        terms: { type: string, minLength: 10, maxLength: 4000 }
        points: { type: integer, minimum: 1, maximum: 1000000 }
        destinationUrl: { type: string, format: uri, maxLength: 2000 }
        category: { type: string, enum: [shopping, subscriptions, surveys, apps, services, finance, travel, food, entertainment, other] }
        requirements: { type: string, minLength: 10, maxLength: 2000 }
        completionSteps: { type: string, minLength: 10, maxLength: 2000 }
        estimatedMinutes: { type: integer, minimum: 1, maximum: 10080 }
        approvalDays: { type: integer, minimum: 1, maximum: 365 }
        expiresAt: { type: [string, "null"], format: date-time }
    V2Offer:
      type: object
      required: [id, partnerId, partnerName, partnerLogoUrl, title, description, terms, points, status, createdAt, destinationUrl, category, requirements, completionSteps, estimatedMinutes, approvalDays, expiresAt]
      properties:
        id: { type: string, format: uuid }
        partnerId: { type: string }
        partnerName: { type: string }
        title: { type: string }
        description: { type: string }
        terms: { type: string }
        points: { type: integer }
        status: { type: string, enum: [draft, approved, rejected] }
        createdAt: { type: string, format: date-time }
        destinationUrl: { type: string, format: uri }
        partnerLogoUrl: { type: [string, "null"], format: uri }
        category: { type: string }
        requirements: { type: string }
        completionSteps: { type: string }
        estimatedMinutes: { type: integer }
        approvalDays: { type: integer }
        expiresAt: { type: [string, "null"], format: date-time }
    V2OfferPage:
      type: object
      required: [items, earnEnabled]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/V2Offer" }
        earnEnabled: { type: boolean }
    V2OfferDetail:
      type: object
      required: [offer, earnEnabled]
      properties:
        offer: { $ref: "#/components/schemas/V2Offer" }
        earnEnabled: { type: boolean }
    V2OfferStartInput:
      type: object
      additionalProperties: false
      required: [idempotencyKey]
      properties:
        idempotencyKey: { type: string, format: uuid }
    V2OfferStart:
      type: object
      required: [clickId, redirectUrl]
      properties:
        clickId: { type: string, format: uuid }
        redirectUrl: { type: string, format: uri }
    V2OfferReviewInput:
      type: object
      additionalProperties: false
      required: [status, reason]
      properties:
        status: { type: string, enum: [approved, rejected] }
        reason: { type: string, minLength: 10, maxLength: 500 }
    V2ConversionReviewInput:
      type: object
      additionalProperties: false
      required: [status, reason]
      properties:
        status: { type: string, enum: [verified, rejected] }
        reason: { type: string, minLength: 10, maxLength: 500 }
    V2ConversionReversalInput:
      type: object
      additionalProperties: false
      required: [reason, idempotencyKey]
      properties:
        reason: { type: string, minLength: 10, maxLength: 500 }
        idempotencyKey: { type: string, pattern: "^[A-Za-z0-9_.:-]{16,128}$" }
    V2ConversionReversalResult:
      type: object
      required: [conversion, compensation, wallet, replayed]
      properties:
        conversion: { $ref: "#/components/schemas/V2Conversion" }
        compensation: { $ref: "#/components/schemas/V2PointsTransaction" }
        wallet: { $ref: "#/components/schemas/V2Wallet" }
        replayed: { type: boolean }
    V2ConversionReversalEvent:
      type: object
      required: [status, reason, actorId, compensationTransactionId, createdAt]
      properties:
        status: { type: string, enum: [reversed] }
        reason: { type: string }
        actorId: { type: string }
        compensationTransactionId: { type: string, format: uuid }
        createdAt: { type: string, format: date-time }
...
    V2Conversion:
      type: object
      required: [id, offerId, offerTitle, points, status, partnerStatus, transactionId, createdAt, updatedAt, accountId, partnerId, eventId, clickId, reversalEvents]
      properties:
        id: { type: string, format: uuid }
        offerId: { type: string, format: uuid }
        offerTitle: { type: string }
        points: { type: integer }
        status: { type: string, enum: [pending, verified, rejected, reversed] }
        partnerStatus: { type: string, enum: [pending, verified, rejected] }
        transactionId: { type: string, format: uuid }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        accountId: { type: string }
        partnerId: { type: string }
        eventId: { type: string }
        clickId: { type: string, format: uuid }
        reversalEvents:
          type: array
          items: { $ref: "#/components/schemas/V2ConversionReversalEvent" }
    V2ConversionPage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/V2Conversion" }
    V2PointsType:
      type: string
      enum: [EARN, REDEEM, REFERRAL, BONUS, ADJUSTMENT, REFUND, REVERSAL, EXPIRATION]
    V2PointsStatus:
      type: string
      enum: [pending, approved, rejected, reversed]
    V2Wallet:
      type: object
      required: [accountId, available, pending, reserved, balance, lifetimeEarned, lifetimeRedeemed]
      properties:
        accountId: { type: string }
        available: { type: integer, minimum: 0 }
        pending: { type: integer, minimum: 0 }
        reserved: { type: integer, minimum: 0 }
        balance: { type: integer, minimum: 0 }
        lifetimeEarned: { type: integer, minimum: 0 }
        lifetimeRedeemed: { type: integer, minimum: 0 }
    V2PointsTransaction:
      type: object
      required: [id, sequence, accountId, type, amount, status, source, reference, description, reason, relatedTransactionId, actorId, createdAt, updatedAt, events]
      properties:
        id: { type: string, format: uuid }
        sequence: { type: string }
        accountId: { type: string }
        type: { $ref: "#/components/schemas/V2PointsType" }
        amount: { type: integer }
        status: { $ref: "#/components/schemas/V2PointsStatus" }
        source: { type: string }
        reference: { type: string }
        description: { type: string }
        reason: { type: string }
        relatedTransactionId: { type: [string, "null"] }
        actorId: { type: string }
        createdAt: { type: string }
        updatedAt: { type: string }
        events:
          type: array
          items: { $ref: "#/components/schemas/V2PointsEvent" }
    V2PointsEvent:
      type: object
      required: [status, delta, reservedDelta, reason, actorId, createdAt]
      properties:
        status: { $ref: "#/components/schemas/V2PointsStatus" }
        delta: { type: integer }
        reservedDelta: { type: integer }
        reason: { type: string }
        actorId: { type: string }
        createdAt: { type: string }
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...