# lib/api-zod/src/generated/types/v2AdminRejectInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API for the BONUSPLAY demo rewards platform
 * OpenAPI spec version: 0.1.0
 */

export interface V2AdminRejectInput {
  /**
     * @minLength 10
     * @maxLength 500
     */
  note: string;
}