# lib/api-zod/src/generated/types/v2AdminRequestInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API for the BONUSPLAY demo rewards platform
 * OpenAPI spec version: 0.1.0
 */
import type { V2AdminRequestInputAction } from './v2AdminRequestInputAction';
import type { V2AdminRequestInputPayload } from './v2AdminRequestInputPayload';

export interface V2AdminRequestInput {
  action: V2AdminRequestInputAction;
  payload: V2AdminRequestInputPayload;
  /**
     * @minLength 10
     * @maxLength 500
     */
  reason: string;
  /**
     * @minLength 16
     * @maxLength 128
     * @pattern ^[A-Za-z0-9_.:-]{16,128}$
     */
  requestKey: string;
}