# lib/api-zod/src/generated/types/v2AuditEntry.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API for the BONUSPLAY demo rewards platform
 * OpenAPI spec version: 0.1.0
 */
import type { V2AuditEntryMetadata } from './v2AuditEntryMetadata';

export interface V2AuditEntry {
  id: string;
  actorId: string;
  actorRole: string;
  action: string;
  entityType: string;
  entityId: string;
  metadata: V2AuditEntryMetadata;
  createdAt: string;
}