# lib/api-zod/src/generated/types/v2AuditPage.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API for the BONUSPLAY demo rewards platform
 * OpenAPI spec version: 0.1.0
 */
import type { V2AuditEntry } from './v2AuditEntry';

export interface V2AuditPage {
  items: V2AuditEntry[];
  /** @nullable */
  nextBefore: string | null;
}