# lib/api-zod/src/generated/types/v2Liability.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API for the BONUSPLAY demo rewards platform
 * OpenAPI spec version: 0.1.0
 */

export interface V2Liability {
  balancePoints: number;
  reservedPoints: number;
  availablePoints: number;
  pendingCreditPoints: number;
  balanceNok: number;
  pendingCreditNok: number;
}